+++ src/cacheout/__pkg__.py
@@ -4,7 +4,7 @@
 __package_name__ = 'cacheout'
 __description__ = 'A caching library for Python'
 __url__ = 'https://github.com/dgilland/cacheout'
-__version__ = '0.6.0'
+__version__ = '0.7.0'
 __author__ = 'Derrick Gilland'
 __email__ = 'dgilland@gmail.com'
 __license__ = 'MIT License'